Drop fixes for bug 1113723
The fixes for bug 1113723 gave sshd more potential access than required
to access Nova's ssh files. As of the RHEL / CentOS 7.3 SELinux base
policy, the following line is present in the SELinux file context set:

/var/lib/[^/]+/\.ssh(/.*)?  all files  system_u:object_r:ssh_home_t:s0

This removes the need for the fix for bug 1113723 by correcting the
context of /var/lib/nova/.ssh/* to ssh_home_t, which sshd can
already access.

Resolves: rhbz#1211734

Signed-off-by: Lon Hohberger <lhh@redhat.com>