local_settings: Set all booleans in on-disk policy
In chroots (for example, when building images using
diskimage-builder or other utilities), access to the
selinux policy in-kernel policy is not allowed. So,
setsebool/getsebool do not work.

Instead, use semanage to modify the on-disk policy
in two passes: Pass (1): install our modules, change
file contexts, and set booleans that already exist
on the systems.  Pass(2): set booleans that were
added in our newly-installed modules from step 1.

Related: rhbz#1489863

Signed-off-by: Lon Hohberger <lhh@redhat.com>