local_settings: Use setfiles
When building images, restorecon is not run since it
requires access to the SELinux policy store in the
host kernel and we do not reload the policy when
building images (since RPM installation is done in
a chroot environment).

Using setfiles works without the need for the
policy to be loaded.

Thanks to Petr Lautrbach for the pointer.

Signed-off-by: Lon Hohberger <lhh@redhat.com>