local_settings.sh.in: add container_file_t context to /var/lib/tripleo-config
We have been managing the container_file_t context for
/var/lib/tripleo-config directory via Ansible until now; but this is an
expensive and risky operation as we rely on another framework.

It would be much simpler for to just use openstack-selinux like we
already do for other directories (e.g. /var/lib/config-data); so this
patch aims to do it.

/var/lib/tripleo-config is a directory which contains tripleo containers
configuration files and it needs this context to operate when SElinux is
enforcing.