Attempt to be smarter with fcontext during updates
fcontext is a slow process, we shouldn't re-apply everything without any
check on the existing rules.
This patch attempts to introduce some smart fcontext management,
checking what we have and what we want.

All the magic is possible thanks to associative arrays, and some string
manipulations on `semanage fcontext -Cl' output. This command gets all
the custom rules present on the system - we therefore just need to do
some filtering in order to keep the rules we actually want.

This patch also avoid some useless repetition - since we're using the
same file list in multiple method, this allows to avoid some mistakes,
wrong copy-paste, or just a "woops, I didn't see I had to add this file
in two places".

On a performance side, this won't change anything for the first install,
but it will make updates faster, since only the needed fcontext calls
will be done.