Allow neutron-keepalived-state-change cache_home_t
The neutron-keepalived-state-change process is
started by the neutron-l3-agent and running as root.

It in turn is using stevedore [1] to load drivers
that has a cache mechanism trying to write to
~/.cache/python-entrypoints by default [2].

This adds a allow for neutron_t processes to access
cache_home_t (~/.cache) that is the default for
stevedore.

This doesn't cause any issues per se since stevedore
passes any IOError on write and just continues so
this is up to our discretion to allow.

[1] https://pypi.org/project/stevedore/
[2] https://github.com/openstack/stevedore/blob/master/stevedore/_cache.py#L47