Move the dac_override rules behind booleans.
As a preparation step for bug 2000945 which aims to reduce dac_override
usage, move the remaining dac_override rules behind booleans.

This patch is currently a noop as the booleans are then enabled. The
end goal is to disable them all once the code fixes are ready and
merged.

Related to rhbz#2000945.