Allow container_t to manage container_log_t files
A new patch in container-selinux[1] adds a new fcontext for
/var/log/containers, setting it as "container_log_t".

Since it's managed by a package, we don't really want to fight with
context.

While we could ensure every container are mounting this location with
the ":z" flag, leading to a relabelling, it's easier and better to just
allow container_t to manage files with this context.

Therefore, this patch does two things:
- allow container_t on container_log_t for files and directories
- remove the specific fcontext for /var/log/containers location

Notes:
- the "test" log is from an enforcing environment, but we can deduce
  that other file management related capabilities (create, unlink and so
  on) are also needed, since it's a log location.
- logrotate_t doesn't need any update: it's running in a container, and
  since container_log_t is of "logfile" type, logrotate can already
  access it.

[1] https://github.com/containers/container-selinux/commit/7e5f3cae10e2d805821fb84dff7418b9e3b0cc1f

Resolves: rhbz#2020210