Add a new option -V to verify installed policies
This new parameter will help ensuring the package did properly install.
It will ensure we're on an SELinux enabled, Enforcing system, then loop
on the different $MODULES to ensure they are present on the system.

In the end, this will help ensuring the package is properly installed,
avoiding future hide'n'seek parties when we're seeing any weird SELinux
issues within TripleO.