Use stderr for error in verification function
`rpm -V' filters the output and displays only stderr content, so we'd be
missing the important bits of the verification.

With this patch in, the command outputs the list of missing modules, the
amount of them, and the final, default rpm verification failure message.

This patch is a follow-up of
https://github.com/redhat-openstack/openstack-selinux/pull/93