[os-virt] allow virtlogd to access /var/lib/nova
As part of https://issues.redhat.com//browse/OSPRH-960
libvirt will be installed on the host not in a container.

libvirt will be used to execute qemu instances with console
files located in /var/lib/nova. As nova is executed in a contaienr
/var/lib/nova is labled with container_file_t

To enable this we need to allow virtlogd to create, delete and
update files in /var/lib/nova

Related: OSPRH-960